* adjust to sum return agregate being float (new version of v-calculate client library) * extra readme info * multiprocessing usage for creation of InfluxDB Point objects

+++ src/api/business_logic/models/aggregate.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass
 @dataclass
 class Aggregate:
     avg: float
-    sum: int
+    sum: float
 
     @classmethod
     def create(cls, values_dict):
